+++ tests/core/serialize/clause/union.cpp
@@ -32,17 +32,25 @@ int main(int, char*[]) {
 
   // Using member function
   SQLPP_COMPARE(select(t.id).from(t).union_all(select(f.id).from(f)),
-                "SELECT tab_bar.id FROM tab_bar UNION ALL SELECT "
-                "tab_foo.id FROM tab_foo");
+                "SELECT tab_bar.id FROM tab_bar "
+                "UNION ALL "
+                "SELECT tab_foo.id FROM tab_foo");
+
+  SQLPP_COMPARE(select(t.id).from(t).union_all(select(f.id).from(f)).order_by(t.id.asc()),
+                "SELECT tab_bar.id FROM tab_bar "
+                "UNION ALL "
+                "SELECT tab_foo.id FROM tab_foo ORDER BY id ASC");
 
   SQLPP_COMPARE(select(t.id).from(t).union_distinct(select(f.id).from(f)),
-                "SELECT tab_bar.id FROM tab_bar UNION DISTINCT SELECT "
-                "tab_foo.id FROM tab_foo");
+                "SELECT tab_bar.id FROM tab_bar "
+                "UNION DISTINCT "
+                "SELECT tab_foo.id FROM tab_foo");
 
   SQLPP_COMPARE(
       select(t.id).from(t).union_distinct(dynamic(true, select(f.id).from(f))),
-      "SELECT tab_bar.id FROM tab_bar UNION DISTINCT SELECT "
-      "tab_foo.id FROM tab_foo");
+      "SELECT tab_bar.id FROM tab_bar "
+      "UNION DISTINCT "
+      "SELECT tab_foo.id FROM tab_foo");
 
   SQLPP_COMPARE(
       select(t.id).from(t).union_distinct(dynamic(false, select(f.id).from(f))),
@@ -52,24 +60,30 @@ int main(int, char*[]) {
                     .from(t)
                     .union_distinct(select(f.id).from(f))
                     .union_all(select(t.id).from(t)),
-                "SELECT tab_bar.int_n AS id FROM tab_bar UNION "
-                "DISTINCT SELECT tab_foo.id FROM tab_foo UNION ALL "
+                "SELECT tab_bar.int_n AS id FROM tab_bar "
+                "UNION DISTINCT "
+                "SELECT tab_foo.id FROM tab_foo "
+                "UNION ALL "
                 "SELECT tab_bar.id FROM tab_bar");
 
   // Using free function
   SQLPP_COMPARE(union_all(select(t.id).from(t), select(f.id).from(f)),
-                "SELECT tab_bar.id FROM tab_bar UNION ALL SELECT "
-                "tab_foo.id FROM tab_foo");
+                "SELECT tab_bar.id FROM tab_bar "
+                "UNION ALL "
+                "SELECT tab_foo.id FROM tab_foo");
 
   SQLPP_COMPARE(union_distinct(select(t.id).from(t), select(f.id).from(f)),
-                "SELECT tab_bar.id FROM tab_bar UNION DISTINCT SELECT "
-                "tab_foo.id FROM tab_foo");
+                "SELECT tab_bar.id FROM tab_bar "
+                "UNION DISTINCT "
+                "SELECT tab_foo.id FROM tab_foo");
 
   SQLPP_COMPARE(union_all(union_distinct(select(t.intN.as(f.id)).from(t),
                                          select(f.id).from(f)),
                           select(t.id).from(t)),
-                "SELECT tab_bar.int_n AS id FROM tab_bar UNION DISTINCT SELECT "
-                "tab_foo.id FROM tab_foo UNION ALL "
+                "SELECT tab_bar.int_n AS id FROM tab_bar "
+                "UNION DISTINCT "
+                "SELECT tab_foo.id FROM tab_foo "
+                "UNION ALL "
                 "SELECT tab_bar.id FROM tab_bar");
 
   return 0;