+++ docs/change_log.md
@@ -16,6 +16,9 @@
     - date_time -> timestamp
   - documented the format of the custom types file
   - recognize and process "ALTER TABLE...ALTER COLUMN...SET DEFAULT" commands, which improves processing of pg_dump output scripts.
+  - forbid order_by, limit, offset, and for_update in union arguments (while mysql and postgresql would allow order_by, limit, and offset if the arguments and enclosed in parentheses, these clauses are not allowed in sqlite3 and parentheses aren't allowed either)
+  - add order_by, limit, and offset to union expressions, e.g. `lhs.union(rhs).order_by(t.id).limit(10).offset(10)`
+  - forbid expressions that require tables in limit and offset, e.g. `limit(t.id)`
 
 ## 0.67